requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
binance-futures-connector==3.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
contourpy==1.0.5
cryptography==38.0.1
cx-Freeze==6.12.0
cycler==0.11.0
dateparser==1.1.1
fonttools==4.38.0
frozenlist==1.3.1
huobi-client @ git+https://github.com/HuobiRDCenter/huobi_Python.git@c75a7fa8b31e99ffc1c173d74dcfcad83682e943
huobipy==1.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
kiwisolver==1.4.4
kucoin-futures-python==1.0.6
matplotlib==3.6.1
multidict==6.0.2
numpy==1.23.4
packaging==21.3
patchelf==0.15.0.0
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybit==2.4.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PySide2==5.15.2.1
PySide6==6.2.0
PySide6-Addons==6.2.0
PySide6-Essentials==6.2
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
service-identity==21.1.0
shiboken2==5.15.2.1
shiboken6==6.2.0
six==1.16.0
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websocket-client==1.4.1
websockets==10.3
yarl==1.8.1
zipp==3.9.0
zope.interface==5.5.0
